Organize customer utils
